Feat/apply permissions

+++ app/Http/Controllers/CondominioController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Helpers\CNPJValidator;
+use App\Helpers\HasPermissions;
 use App\Repository\Eloquent\CondominioRepository;
 use App\Interfaces\DeleteInterface;
 use App\Interfaces\GetAllInterface;
@@ -21,6 +22,7 @@ class CondominioController extends Controller implements GetAllInterface,
     private $repository;
     private $validator;
     private $cnpjValidator;
+    private $hasPermissions;
 
     private $rules = [
         'name' => 'required|string|max:255',
@@ -36,23 +38,41 @@ class CondominioController extends Controller implements GetAllInterface,
         'taxabasicaagua'=> 'required|numeric',
     ];
 
-    public function __construct(CondominioRepository $repository, Validator $validator, CNPJValidator $cnpjValidator) {
+    public function __construct(
+        CondominioRepository $repository, 
+        Validator $validator, 
+        CNPJValidator $cnpjValidator,
+        HasPermissions $hasPermissions
+    ) {
         $this->repository = $repository;
         $this->validator = $validator;
         $this->cnpjValidator = $cnpjValidator;
+        $this->hasPermissions = $hasPermissions;
     }
 
-    public function getAll(): JsonResponse {
+    public function getAll(Request $request): JsonResponse {
+        if (!$this->hasPermissions->hasPermission($request->user()->perfil_id, 'CAD_CONDOMINIO', 'consultar')) {
+            return response()->json('Not Found', 404);
+        }
+
         $condominios = $this->repository->getAll();
         return response()->json($condominios);
     }
 
-    public function getById($id): JsonResponse {
+    public function getById(Request $request, $id): JsonResponse {
+        if (!$this->hasPermissions->hasPermission($request->user()->perfil_id, 'CAD_CONDOMINIO', 'consultar')) {
+            return response()->json('Not Found', 404);
+        }
+
         $condominio = $this->repository->getById($id);
         return response()->json($condominio);
     }
 
     public function save(Request $request): JsonResponse {
+        if (!$this->hasPermissions->hasPermission($request->user()->perfil_id, 'CAD_CONDOMINIO', 'inserir')) {
+            return response()->json('Not Found', 404);
+        }
+
         $data = $request->all();
         $this->validateFields($data, $this->rules);
 
@@ -66,6 +86,10 @@ class CondominioController extends Controller implements GetAllInterface,
     }
 
     public function update(Request $request, $id): JsonResponse {
+        if (!$this->hasPermissions->hasPermission($request->user()->perfil_id, 'CAD_CONDOMINIO', 'alterar')) {
+            return response()->json('Not Found', 404);
+        }
+
         $data = $request->all();
         $this->validateFields($data, $this->rules);
 
@@ -79,6 +103,10 @@ class CondominioController extends Controller implements GetAllInterface,
     }
 
     public function delete(Request $request, $id): JsonResponse {
+        if (!$this->hasPermissions->hasPermission($request->user()->perfil_id, 'CAD_CONDOMINIO', 'excluir')) {
+            return response()->json('Not Found', 404);
+        }
+
         $condominio = $this->repository->delete($id);
         if ($condominio) {
             return response()->json($condominio);
